+++ packages/components/pages/f-mfa/src/components/Help.vue
@@ -29,7 +29,10 @@
                     <i18n
                         path="helpInfo.instructionsPoint4Text"
                         tag="li">
-                        <a :href="$t('helpInfo.instructionsHelpLink')"><strong>{{ $t('helpInfo.instructionsHelpText') }}</strong></a>
+                        <a
+                            data-test-id="f-mfa-help-faq-link"
+                            :href="$t('helpInfo.instructionsHelpLink')">
+                            <strong>{{ $t('helpInfo.instructionsHelpText') }}</strong></a>
                     </i18n>
                 </ul>
 
@@ -48,7 +51,8 @@
                     button-type="link"
                     :href="loginLinkWithReturnUrl"
                     button-size="large"
-                    data-test-id="mfa-help-login-link">
+                    data-test-id="mfa-help-login-link"
+                    @click="recordAnalytics">
                     {{ $t('helpInfo.secondaryButtonText') }}
                 </f-button>
             </div>
@@ -68,6 +72,10 @@ import {
     BagSadBgIcon
 } from '@justeat/f-vue-icons';
 import tenantConfigs from '../tenants';
+import {
+    buildEvent,
+    HELP_LOGIN
+} from '../services/EventBuilder';
 
 export default {
     name: 'VMfa',
@@ -100,6 +108,12 @@ export default {
     computed: {
         loginLinkWithReturnUrl () {
             return `/account/login?returnUrl=${encodeURIComponent(this.returnUrl)}`;
+        }
+    },
+
+    methods: {
+        recordAnalytics () {
+            this.$gtm.pushEvent(buildEvent(HELP_LOGIN));
         }
     }
 };