+++ packages/components/pages/f-mfa/CHANGELOG.md
@@ -4,6 +4,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.10.0
+------------------------------
+*August 17, 2022*
+
+### Changed
+- Changed mfa load success log message to `Info`.
+- Help screen primary button text.
+  
+### Removed
+- Mfa Form field validation/message.
+
+### Added
+- GTM events.
+- Back button to error screen.
+
+
 v0.9.1
 ------------------------------
 *August 11, 2022*